renamed the project and package names
